+++ src/IncludedComponents/views/administration/dashboard.blade.php
@@ -70,6 +70,24 @@
                     </a>
                 </div>
             <?php } ?>
+            <?php if($opt["user"]->checkPermission("admin.organizations.list")) { ?>
+                <div class="col-12 col-sm-6 col-md-3 mb-3">
+                    <a href="<?= "$opt[root]z/organizations" ?>" data-test="dash-organizations" class="card shadow-sm h-100 text-decoration-none text-muted">
+                        <div class="card-body d-flex align-items-center justify-content-center py-5">
+                            <i class="fa fa-building fa-4x"></i>
+                        </div>
+                    </a>
+                </div>
+            <?php } ?>
+            <?php if($opt["user"]->checkPermission("admin.organizations.create")) { ?>
+                <div class="col-12 col-sm-6 col-md-3 mb-3">
+                    <a href="<?= "$opt[root]z/add_organization" ?>" data-test="dash-add-organization" class="card shadow-sm h-100 text-decoration-none text-muted">
+                        <div class="card-body d-flex align-items-center justify-content-center py-5">
+                            <i class="fa fa-plus-circle fa-4x"></i>
+                        </div>
+                    </a>
+                </div>
+            <?php } ?>
         </div>
 
         <h2 class="mt-4 mb-2 pl-1 font-weight-bold h5">